Add Order Services integration tests
