+++ src/test/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringdatastoreapi/integration/resources/OrderServicesControllerIntegrationTest.kt
@@ -1,0 +1,47 @@
+package uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.integration.resources
+
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+import org.springframework.test.context.ActiveProfiles
+import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.mocks.MockEmDatastoreClient
+
+@ActiveProfiles("integration")
+class OrderServicesControllerIntegrationTest : ControllerIntegrationBase() {
+  @Nested
+  @DisplayName("GET /orders/{orderId}/services")
+  inner class GetMonitoringEvents {
+    @BeforeEach
+    fun setup() {
+      MockEmDatastoreClient.Companion.addResponseFile("successfulOrderServicesListResponse")
+    }
+
+    @Test
+    fun `should return 401 unauthorized if no authorization header`() {
+      noAuthHeaderRespondsWithUnauthorizedTest("/orders/234/services")
+    }
+
+    @Test
+    fun `should return 403 forbidden if no role in authorization header`() {
+      noRoleInAuthHeaderRespondsWithForbiddenTest("/orders/234/services")
+    }
+
+    @Test
+    fun `should return 403 forbidden if wrong role in authorization header`() {
+      wrongRolesRespondsWithForbiddenTest("/orders/234/services", listOf("ROLE_WRONG"))
+    }
+
+    @Test
+    fun `should return OK with valid auth header, role`() {
+      val uri = "/orders/234/services"
+
+      webTestClient.get()
+        .uri(uri)
+        .headers(setAuthorisation())
+        .exchange()
+        .expectStatus()
+        .isOk
+    }
+  }
+}
